Minor fiexs

+++ db/migrate/20190304142110_create_reviews.rb
@@ -2,7 +2,7 @@ class CreateReviews < ActiveRecord::Migration[5.2]
   def change
     create_table :reviews do |t|
       t.text :content
-      t.float :score
+      t.integer :score
       t.references :user, foreign_key: true
       t.references :flat, foreign_key: true
